requirements.txt
requests~=2.32.0
python-dotenv~=1.0.1
shapely~=2.0.6
networkx
osmnx
geojson~=3.1.0
shapely
mapbox-vector-tile
python-dotenv
mongo_manager_juan_palma_borda
